MINOR :: Changed u64 to usize/isize

+++ efi/src/arch/amd64/mod.rs
@@ -15,7 +15,7 @@ pub fn halt() {
     unsafe { asm!("hlt"); }
 }
 
-pub const R_RELATIVE: u64 = 8;
+pub const R_RELATIVE: usize = 8;
 
 #[inline(always)]
 pub fn stack_ptr() -> usize {
