+++ kernel/src/kargs.rs
@@ -33,9 +33,9 @@ pub struct SysInfo {
 
 #[repr(C)]
 pub struct RelaEntry {
-    pub offset: u64,
-    pub info: u64,
-    pub addend: u64
+    pub offset: usize,
+    pub info: usize,
+    pub addend: isize
 }
 
 #[repr(C)]
